db: Parse leases *after* dropping privileges.
Prior to this commit, the leases file was parsed while the process
was running as root prior to entering the pledge / capsicum sandbox.

This was problematic because the lease file contains untrusted input
that passes through very complicated parsing logic. If an attacker
achieved arbitrary code execution through this logic, their code
would run as root.